+++ src/App.js
@@ -2,10 +2,11 @@ import React, { Component } from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
 import Auth from './components/auth';
 import './styles/reset.css';
+import './styles/app.css';
 import { Provider } from 'react-redux';
 import storeFactory from './lib/store';
-import Home from './home/home';
-import Nav from './header/nav';
+import Home from './components/home/home';
+import Nav from './components/header/nav';
 import CreateGame from './components/create-game';
 const store = storeFactory();
 
@@ -28,6 +29,9 @@ class App extends Component {
                 <Route exact path='/auth/:type' component={Auth}/>
               </div>
             </main>
+            <footer>
+              &copy; DeltaV
+            </footer>
           </div>
         </BrowserRouter>
       </Provider>
