+++ src/components/header/nav.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import Modal from './modal';
-import Auth from '../components/auth';
+import Auth from '../auth';
+import './nav.css';
 
 export default class Nav extends Component {
   constructor(props) {
