+++ src/components/home/home.css
@@ -1,0 +1,12 @@
+html {
+  height: 100%;
+}
+
+.homeBody {
+  height: 100%;
+  background-color: rgb(0, 62, 70);
+}
+
+h2 {
+  font-size: 150%;
+}
