+++ src/styles/app.css
@@ -1,0 +1,13 @@
+body {
+  height: 100%;
+  bottom: 0;
+}
+
+.layout {
+  height: 100%;
+}
+
+footer {
+  position: fixed;
+  bottom: 0;
+}
